fix: revert V4 spice_level to SMALLINT (checksum mismatch on prod)
V4 was already applied to prod with the SMALLINT declaration. Changing
V4 to INT triggered a Flyway checksum mismatch on the next deploy
(Applied: 1530347634 vs Resolved: -1598023092). Instead, keep V4
as-is and let V5 widen to INTEGER for prod. Test container
applies V4 (SMALLINT) and V5 (widen) cleanly.

The earlier 'revert and try again' message was wrong direction;
the V4 content is back to the original SMALLINT now.

+++ src/main/resources/db/migration/V4__product_metadata.sql
@@ -15,7 +15,7 @@ ALTER TABLE product
     ADD COLUMN weight_grams  INT;
 
 ALTER TABLE menu_item
-    ADD COLUMN spice_level   INT;
+    ADD COLUMN spice_level   SMALLINT;
 
 ALTER TABLE product
     ADD CONSTRAINT ck_product_weight_grams
